+++ concatenate.js
@@ -1,4 +1,3 @@
-const fs = require('fs-extra');
 const concat = require('concat');
 (async function build() {
   const es5Files = [
@@ -14,29 +13,6 @@ const concat = require('concat');
     './dist/lforms/webcomponent/main-es2015.js',
   ]
 
-  const copiedFiles = [
-    'runtime-es5.js',
-    'runtime-es5.js.map',
-    'runtime-es2015.js',
-    'runtime-es2015.js.map',
-    'polyfills-es5.js',
-    'polyfills-es5.js.map',
-    'polyfills-es2015.js',
-    'polyfills-es2015.js.map',
-    'main-es5.js',
-    'main-es5.js.map',
-    'main-es2015.js',
-    'main-es2015.js.map',
-    'scripts.js',
-    'styles.css',
-    'magnifying_glass.png',
-    'down_arrow_gray_10_10.png'
-  ]
-  // await fs.ensureDir('dist/lforms/webcomponent')
   await concat(es5Files, 'dist/lforms/webcomponent/lhc-forms.es5.js');
   await concat(es2015Files, 'dist/lforms/webcomponent/lhc-forms.es2015.js');
-  // copiedFiles.forEach(async (file) => {
-  //   await fs.copyFile('./dist/lforms/' + file, 'dist/lforms/webcomponent/' + file);
-  // })
-  // await fs.copy('./dist/lforms/assets/', 'dist/lforms/webcomponent/assets/' )
 })()