+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 - Updated angular from v12 to v14
 ### Changed
 - Removed the support for IE 11.
-- Only ES2020 files (no ES5 files) are in the build files.
+- Only ES2017 files (no ES5 files) are in the build files.
   The concatnated js file name is now lhc-forms.js, not
   lhc-forms.es5.js or lhc-forms.es2015.js